Fix bug #3002075 - value temperature is missing
udisks plugin was using smart time parameter as a proxy for testing if SMART
data is available for a drive. Since at start-up this is 0, this was causing
it not to detect drives which have previously been detected. Instead actually
check SMART is supported instead. We then need to check that smart time is
non-zero when querying SMART data otherwise we get back garbage if it is 0.